refactor: move SemesterProject to Project mocks
